refactor(012): make event-coordinator a pure event router (US3 T042)
Move the login-handler cache-rebuild IO into StorageService.refreshCacheOnLogin
and the post-login table upgrade into enhancers/table-upgrade.ts. The coordinator
(216 lines) now only routes events to those collaborators and the shared reveal.
